pyproject.toml
[project]
name = "llm-tools"
version = "0.0.1"
description = "Add your description here"
authors = [
    { name = "Artem Durynin", email = "artem5542@gmail.com" }
]
dependencies = [
    "vllm>=0.8",
    "pillow>=11.0.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
    "openai>=1.73.0",
    "rouge-score>=0.1.2",
    "ruff>=0.11.5",
    "transformers>=4.51.2",
    "pyarrow>=19.0.1",
    "peft>=0.15.1",
    "evaluate>=0.4.3",
    "pytest>=8.3.5",
    "trl>=0.15.2",
    "unsloth>=2025.4.7",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "cairosvg>=2.7.1",
    "pip>=25.1.1",
    "wandb>=0.19.11",
]

readme = "README.md"
requires-python = ">=3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/llm_tools"]

#[[tool.rye.sources]]
#name = "torch"
#url = "https://download.pytorch.org/whl/cpu"
#type = "index"

[tool.hatch.dependencies.overrides]
# Add Git dependencies as overrides for `pip` installations
unsloth = {url = "git+https://github.com/unslothai/unsloth.git"}
unsloth_zoo = {url = "git+https://github.com/unslothai/unsloth-zoo.git"}
